core_files/requirements.txt
Appium-Python-Client==0.47
behave==1.2.6
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
coverage==4.5.4
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
parse==1.15.0
parse-type==0.5.2
pathvalidate==2.3.0
Paver==1.3.4
pycodestyle==2.5.0
pyflakes==2.1.1
PyYAML==5.3
requests==2.23.0
selenium==3.141.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
webdriver-manager==3.2.1
